fix(ui): load controls left and add basic speaker panel
